+++ modules/tenant/index.ts
@@ -1,6 +1,7 @@
 export { TenantDomainService } from "./services/TenantDomainService";
 export { DnsVerificationService } from "./services/DnsVerificationService";
 export { K8sCertificateService } from "./services/K8sCertificateService";
+export { K8sIngressRouteService } from "./services/K8sIngressRouteService";
 export {
   buildSlugCandidate,
   buildSlugVariant,